requirements.txt
allure-pytest
allure-python-commons
pytest
requests
pytest-xdist